Revert "fix: .env파일을 만들지 않게 변경 (#156)"
This reverts commit 28bf6d9e0d3109dfc20611864dd342ac0ead48f1.
