+++ .github/workflows/ci.yml
@@ -31,15 +31,13 @@ jobs:
 
       - name: Test
         env:
-          MYSQL_USER: ${{ secrets.MYSQL_USER }}
-          MYSQL_PASSWORD: ${{ secrets.MYSQL_PASSWORD }}
-          DATABASE_NAME: ${{ secrets.DATABASE_NAME }}
-          DATABASE_URL: ${{ secrets.DATABASE_URL }}
-          DATABASE_PORT: ${{ secrets.DATABASE_PORT }}
-          JWT_SECRET: ${{ secrets.JWT_SECRET }}
-          GOOGLE_NEARBYSEARCH_API_KEY: ${{ secrets.GOOGLE_NEARBYSEARCH_API_KEY }}
-          GOOGLE_CLIENT_ID: ${{ secrets.GOOGLE_CLIENT_ID }}
-          GOOGLE_SECRET_KEY: ${{ secrets.GOOGLE_SECRET_KEY }}
+          JWT_SECRET: ${{secrets.JWT_SECRET}}
+          GOOGLE_NEARBYSEARCH_API_KEY: ${{secrets.GOOGLE_NEARBYSEARCH_API_KEY}}
+          MYSQL_USER: ${{secrets.MYSQL_USER}}
+          MYSQL_PASSWORD: ${{secrets.MYSQL_PASSWORD}}
+          DATABASE_NAME: ${{secrets.DATABASE_NAME}}
+          DATABASE_URL: ${{secrets.DATABASE_URL}}
+          DATABASE_PORT: ${{secrets.DATABASE_PORT}}
 
         run: yarn test
