+++ src/webpush/webpushController.ts
@@ -1,10 +1,10 @@
 import { Body, Controller, Post, Route, SuccessResponse, Tags } from 'tsoa';
 import { PushSubscription } from 'web-push';
-import { WebpushService } from './webpushService';
+import { webpushService } from './webpushService';
 
 @Route('webpush')
 @Tags('Webpush')
-export class WebpushController extends Controller {
+export class webpushController extends Controller {
   /**
    * subscription으로 대상을 특정해서 web push를 보냅니다
    * @param id 보내줄 Todo의 UUID
@@ -13,7 +13,7 @@ export class WebpushController extends Controller {
   @SuccessResponse('201', 'Created')
   @Post('{id}')
   public async pushTodo(id: string, @Body() reqBody: { subscription: PushSubscription }) {
-    const res = await new WebpushService().pushTodo(id, reqBody.subscription);
+    const res = await new webpushService().pushTodo(id, reqBody.subscription);
     if (!res) {
       this.setStatus(500);
     }
