+++ src/webpush/webpushService.ts
@@ -1,18 +1,7 @@
 import { Todo } from '@/todos/todo';
 import webpush, { PushSubscription } from 'web-push';
 
-export class WebpushService {
-  static publicVapidKey = 'BHCoqzR03UrjuAFGPoTDB5t6o05z5K3EYJ1cuZVj9sPF6FxNsS-b7y4ClNaS11L9EUpmT-wUyeZAivwGbkwMAjY';
-  static privateVapidKey = 'QNapwA1rlszq7UdCqLo5s5aORi5jAAlCEFMEfZaw4tU';
-
-  constructor() {
-    webpush.setVapidDetails(
-      'mailto:leehj0110@kakao.com',
-      WebpushService.publicVapidKey,
-      WebpushService.privateVapidKey,
-    );
-  }
-
+export class webpushService {
   public async pushTodo(id: string, subscription: PushSubscription) {
     const todo = await Todo.findByPk(id);
